requirements.txt
streamlit
pandas
matplotlib
langchain
google-generativeai
langchain-google-genai
python-dotenv